v6.0.0rc1
- Fix imported file path
- Add evaluation scripts
