+++ cjs-eval.js
@@ -1,10 +1,7 @@
 // eval.js
 // to quickly test with a single url or file
 
-const { readFileSync, existsSync } = require('fs')
-
-const isValidUrl = require('./src/utils/isValidUrl')
-const { extract } = require('./index')
+const { extract } = require('./dist/cjs/article-parser.js')
 
 const extractFromUrl = async (url) => {
   try {
@@ -15,21 +12,10 @@ const extractFromUrl = async (url) => {
   }
 }
 
-const extractFromFile = async (fpath) => {
-  try {
-    const html = readFileSync(fpath, 'utf8')
-    const art = await extract(html)
-    console.log(art)
-  } catch (err) {
-    console.trace(err)
-  }
-}
-
 const init = (argv) => {
   if (argv.length === 3) {
     const input = argv[2]
-    const isUrl = isValidUrl(input)
-    return isUrl ? extractFromUrl(input) : existsSync(input) ? extractFromFile(input) : false
+    return extractFromUrl(input)
   }
   return 'Nothing to do!'
 }
